feat: hookups for update fn
